+++ internals-js/src/__tests__/subgraphValidation.test.ts
@@ -640,8 +640,15 @@ describe('custom error message for misnamed directives', () => {
   });
 });
 
-function buildAndValidate(doc: DocumentNode): Subgraph {
-  const name = 'S';
+function buildAndValidate(
+  subgraphDefs: DocumentNode,
+  options?: {
+    subgraphName?: string,
+    asFed2?: boolean,
+  }
+): Subgraph {
+  const doc = (options?.asFed2 ?? true) ? asFed2SubgraphDocument(subgraphDefs) : subgraphDefs;
+  const name = options?.subgraphName ?? 'S';
   return buildSubgraph(name, `http://${name}`, doc).validate();
 }
 
@@ -728,7 +735,7 @@ describe('@core/@link handling', () => {
       }
     `;
 
-    validateFullSchema(buildAndValidate(doc));
+    validateFullSchema(buildAndValidate(doc, { asFed2: false }));
   });
 
   it('expands definitions if both the federation spec and link spec are linked', () => {
@@ -742,11 +749,11 @@ describe('@core/@link handling', () => {
       }
     `;
 
-    validateFullSchema(buildAndValidate(doc));
+    validateFullSchema(buildAndValidate(doc, { asFed2: false }));
   });
 
   it('is valid if a schema is complete from the get-go', () => {
-    validateFullSchema(buildAndValidate(gql(expectedFullSchema)));
+    validateFullSchema(buildAndValidate(gql(expectedFullSchema), { asFed2: false }));
   });
 
   it('expands missing definitions when some are partially provided', () => {
@@ -817,7 +824,7 @@ describe('@core/@link handling', () => {
     // calls the graphQL-js validation, so we can be somewhat sure that if something necessary wasn't expanded
     // properly, we would have an issue. The main reason we did validate the full schema in prior tests is
     // so we had at least one full example of a subgraph expansion in the tests.
-    docs.forEach((doc) => buildAndValidate(doc));
+    docs.forEach((doc) => buildAndValidate(doc, { asFed2: false }));
   });
 
   it('allows known directives with incomplete but compatible definitions', () => {
@@ -893,7 +900,7 @@ describe('@core/@link handling', () => {
     ];
 
     // Like above, we really only care that the examples validate.
-    docs.forEach((doc) => buildAndValidate(doc));
+    docs.forEach((doc) => buildAndValidate(doc, { asFed2: false }));
   });
 
   it('errors on invalid known directive location', () => {
@@ -1025,7 +1032,7 @@ describe('@core/@link handling', () => {
     `;
 
     // Just making sure this don't error out.
-    buildAndValidate(doc);
+    buildAndValidate(doc, { asFed2: false });
   });
 
   it('allows defining a repeatable directive as non-repeatable but validates usages', () => {
@@ -1064,7 +1071,7 @@ describe('federation 1 schema', () => {
       directive @requires on FIELD_DEFINITION
     `;
 
-    buildAndValidate(doc);
+    buildAndValidate(doc, { asFed2: false });
   });
 
   it('accepts federation directive definitions with nullable arguments', () => {
@@ -1086,7 +1093,7 @@ describe('federation 1 schema', () => {
       directive @requires(fields: String) on FIELD_DEFINITION
     `;
 
-    buildAndValidate(doc);
+    buildAndValidate(doc, { asFed2: false });
   });
 
   it('accepts federation directive definitions with "FieldSet" type instead of "_FieldSet"', () => {
@@ -1103,7 +1110,7 @@ describe('federation 1 schema', () => {
       directive @key(fields: FieldSet) on OBJECT | INTERFACE
     `;
 
-    buildAndValidate(doc);
+    buildAndValidate(doc, { asFed2: false });
   });
 
   it('rejects federation directive definition with unknown arguments', () => {
@@ -1128,16 +1135,52 @@ describe('federation 1 schema', () => {
 })
 
 describe('@shareable', () => {
-  it('can only be applied to fields of object types', () => {
+  it('allowed on interface field if all implementation follow suit', () => {
     const doc = gql`
       interface I {
         a: Int @shareable
+      }
+
+      type T1 implements I {
+        a: Int @shareable
+      }
+
+      # Checks that having it in a @key works, since key field are @shareable by default
+      type T2 implements I @key(fields: "a") {
+        a: Int
+      }
+
+      # Checks @external is fine too
+      type T3 implements I {
+        a: Int @external
+      }
+    `;
+
+    buildAndValidate(doc);
+  });
+
+  it('rejects if on interface field but not all implementations', () => {
+    const doc = gql`
+      interface I {
+        a: Int @shareable
+      }
+
+      type T1 implements I {
+        a: Int
+      }
+
+      type T2 implements I {
+        a: Int @shareable
+      }
+
+      type T3 implements I {
+        a: Int
       }
     `;
 
     expect(buildForErrors(doc)).toStrictEqual([[
       'INVALID_SHAREABLE_USAGE',
-      '[S] Invalid use of @shareable on field "I.a": only object type fields can be marked with @shareable'
+      '[S] Interface field "I.a" is marked @shareable but its implementation in types "T1" and "T3" is not (note that if you do not want to force the field to be @shareable in all implementations, you can simply remove @shareable from "I.a")'
     ]]);
   });